+++ src/whatsapp/controllers/instance.controller.ts
@@ -3,7 +3,7 @@ import { isURL } from 'class-validator';
 import EventEmitter2 from 'eventemitter2';
 import { v4 } from 'uuid';
 
-import { ConfigService, HttpServer } from '../../config/env.config';
+import { ConfigService, HttpServer, WaBusiness } from '../../config/env.config';
 import { Logger } from '../../config/logger.config';
 import { BadRequestException, InternalServerErrorException } from '../../exceptions';
 import { RedisCache } from '../../libs/redis.client';
@@ -12,6 +12,7 @@ import { RepositoryBroker } from '../repository/repository.manager';
 import { AuthService, OldToken } from '../services/auth.service';
 import { CacheService } from '../services/cache.service';
 import { ChatwootService } from '../services/chatwoot.service';
+import { IntegrationService } from '../services/integration.service';
 import { WAMonitoringService } from '../services/monitor.service';
 import { RabbitmqService } from '../services/rabbitmq.service';
 import { SettingsService } from '../services/settings.service';
@@ -19,8 +20,9 @@ import { SqsService } from '../services/sqs.service';
 import { TypebotService } from '../services/typebot.service';
 import { WebhookService } from '../services/webhook.service';
 import { WebsocketService } from '../services/websocket.service';
-import { WAStartupService } from '../services/whatsapp.service';
-import { Events, wa } from '../types/wa.types';
+import { BaileysStartupService } from '../services/whatsapp.baileys.service';
+import { BusinessStartupService } from '../services/whatsapp.business.service';
+import { Events, Integration, wa } from '../types/wa.types';
 
 export class InstanceController {
   constructor(
@@ -36,6 +38,7 @@ export class InstanceController {
     private readonly rabbitmqService: RabbitmqService,
     private readonly sqsService: SqsService,
     private readonly typebotService: TypebotService,
+    private readonly integrationService: IntegrationService,
     private readonly cache: RedisCache,
     private readonly chatwootCache: CacheService,
   ) {}
@@ -50,6 +53,7 @@ export class InstanceController {
     events,
     qrcode,
     number,
+    integration,
     token,
     chatwoot_account_id,
     chatwoot_token,
@@ -57,12 +61,16 @@ export class InstanceController {
     chatwoot_sign_msg,
     chatwoot_reopen_conversation,
     chatwoot_conversation_pending,
+    chatwoot_import_contacts,
+    chatwoot_import_messages,
+    chatwoot_days_limit_import_messages,
     reject_call,
     msg_call,
     groups_ignore,
     always_online,
     read_messages,
     read_status,
+    sync_full_history,
     websocket_enabled,
     websocket_events,
     rabbitmq_enabled,
@@ -83,14 +91,31 @@ export class InstanceController {
       this.logger.verbose('checking duplicate token');
       await this.authService.checkDuplicateToken(token);
 
+      if (!token && integration === Integration.WHATSAPP_BUSINESS) {
+        throw new BadRequestException('token is required');
+      }
+
       this.logger.verbose('creating instance');
-      const instance = new WAStartupService(
-        this.configService,
-        this.eventEmitter,
-        this.repository,
-        this.cache,
-        this.chatwootCache,
-      );
+      let instance: BaileysStartupService | BusinessStartupService;
+      if (integration === Integration.WHATSAPP_BUSINESS) {
+        instance = new BusinessStartupService(
+          this.configService,
+          this.eventEmitter,
+          this.repository,
+          this.cache,
+          this.chatwootCache,
+        );
+        await this.waMonitor.saveInstance({ integration, instanceName, token, number });
+      } else {
+        instance = new BaileysStartupService(
+          this.configService,
+          this.eventEmitter,
+          this.repository,
+          this.cache,
+          this.chatwootCache,
+        );
+      }
+
       instance.instanceName = instanceName;
 
       const instanceId = v4();
@@ -147,6 +172,8 @@ export class InstanceController {
               'GROUP_UPDATE',
               'GROUP_PARTICIPANTS_UPDATE',
               'CONNECTION_UPDATE',
+              'LABELS_EDIT',
+              'LABELS_ASSOCIATION',
               'CALL',
               'NEW_JWT_TOKEN',
               'TYPEBOT_START',
@@ -197,6 +224,8 @@ export class InstanceController {
               'GROUP_UPDATE',
               'GROUP_PARTICIPANTS_UPDATE',
               'CONNECTION_UPDATE',
+              'LABELS_EDIT',
+              'LABELS_ASSOCIATION',
               'CALL',
               'NEW_JWT_TOKEN',
               'TYPEBOT_START',
@@ -244,6 +273,8 @@ export class InstanceController {
               'GROUP_UPDATE',
               'GROUP_PARTICIPANTS_UPDATE',
               'CONNECTION_UPDATE',
+              'LABELS_EDIT',
+              'LABELS_ASSOCIATION',
               'CALL',
               'NEW_JWT_TOKEN',
               'TYPEBOT_START',
@@ -291,6 +322,8 @@ export class InstanceController {
               'GROUP_UPDATE',
               'GROUP_PARTICIPANTS_UPDATE',
               'CONNECTION_UPDATE',
+              'LABELS_EDIT',
+              'LABELS_ASSOCIATION',
               'CALL',
               'NEW_JWT_TOKEN',
               'TYPEBOT_START',
@@ -342,12 +375,30 @@ export class InstanceController {
         always_online: always_online || false,
         read_messages: read_messages || false,
         read_status: read_status || false,
+        sync_full_history: sync_full_history ?? false,
       };
 
       this.logger.verbose('settings: ' + JSON.stringify(settings));
 
       this.settingsService.create(instance, settings);
 
+      let webhook_wa_business = null,
+        access_token_wa_business = '';
+
+      if (integration === Integration.WHATSAPP_BUSINESS) {
+        if (!number) {
+          throw new BadRequestException('number is required');
+        }
+        const urlServer = this.configService.get<HttpServer>('SERVER').URL;
+        webhook_wa_business = `${urlServer}/webhook/whatsapp/${encodeURIComponent(instance.instanceName)}`;
+        access_token_wa_business = this.configService.get<WaBusiness>('WA_BUSINESS').TOKEN_WEBHOOK;
+      }
+
+      this.integrationService.create(instance, {
+        integration,
+        number,
+        token,
+      });
       if (!chatwoot_account_id || !chatwoot_token || !chatwoot_url) {
         let getQrcode: wa.QrCode;
 
@@ -362,6 +413,9 @@ export class InstanceController {
           instance: {
             instanceName: instance.instanceName,
             instanceId: instanceId,
+            integration: integration,
+            webhook_wa_business,
+            access_token_wa_business,
             status: 'created',
           },
           hash,
@@ -444,6 +498,9 @@ export class InstanceController {
           number,
           reopen_conversation: chatwoot_reopen_conversation || false,
           conversation_pending: chatwoot_conversation_pending || false,
+          import_contacts: chatwoot_import_contacts ?? true,
+          import_messages: chatwoot_import_messages ?? true,
+          days_limit_import_messages: chatwoot_days_limit_import_messages ?? 60,
           auto_create: true,
         });
       } catch (error) {
@@ -454,6 +511,9 @@ export class InstanceController {
         instance: {
           instanceName: instance.instanceName,
           instanceId: instanceId,
+          integration: integration,
+          webhook_wa_business,
+          access_token_wa_business,
           status: 'created',
         },
         hash,
@@ -494,6 +554,9 @@ export class InstanceController {
           sign_msg: chatwoot_sign_msg || false,
           reopen_conversation: chatwoot_reopen_conversation || false,
           conversation_pending: chatwoot_conversation_pending || false,
+          import_contacts: chatwoot_import_contacts ?? true,
+          import_messages: chatwoot_import_messages ?? true,
+          days_limit_import_messages: chatwoot_days_limit_import_messages || 60,
           number,
           name_inbox: instance.instanceName,
           webhook_url: `${urlServer}/chatwoot/webhook/${encodeURIComponent(instance.instanceName)}`,
@@ -579,13 +642,13 @@ export class InstanceController {
     };
   }
 
-  public async fetchInstances({ instanceName, instanceId }: InstanceDto) {
+  public async fetchInstances({ instanceName, instanceId, number }: InstanceDto) {
     if (instanceName) {
       this.logger.verbose('requested fetchInstances from ' + instanceName + ' instance');
       this.logger.verbose('instanceName: ' + instanceName);
       return this.waMonitor.instanceInfo(instanceName);
-    } else if (instanceId) {
-      return this.waMonitor.instanceInfoById(instanceId);
+    } else if (instanceId || number) {
+      return this.waMonitor.instanceInfoById(instanceId, number);
     }
 
     this.logger.verbose('requested fetchInstances (all instances)');
@@ -601,11 +664,7 @@ export class InstanceController {
     }
 
     try {
-      this.logger.verbose('logging out instance: ' + instanceName);
-      await this.waMonitor.waInstances[instanceName]?.client?.logout('Log out instance: ' + instanceName);
-
-      this.logger.verbose('close connection instance: ' + instanceName);
-      this.waMonitor.waInstances[instanceName]?.client?.ws?.close();
+      this.waMonitor.waInstances[instanceName]?.logoutInstance();
 
       return { status: 'SUCCESS', error: false, response: { message: 'Instance logged out' } };
     } catch (error) {