+++ src/whatsapp/whatsapp.module.ts
@@ -9,6 +9,7 @@ import { ChatController } from './controllers/chat.controller';
 import { ChatwootController } from './controllers/chatwoot.controller';
 import { GroupController } from './controllers/group.controller';
 import { InstanceController } from './controllers/instance.controller';
+import { LabelController } from './controllers/label.controller';
 import { ProxyController } from './controllers/proxy.controller';
 import { RabbitmqController } from './controllers/rabbitmq.controller';
 import { SendMessageController } from './controllers/sendMessage.controller';
@@ -23,6 +24,7 @@ import {
   ChatModel,
   ChatwootModel,
   ContactModel,
+  IntegrationModel,
   MessageModel,
   MessageUpModel,
   ProxyModel,
@@ -33,11 +35,14 @@ import {
   WebhookModel,
   WebsocketModel,
 } from './models';
+import { LabelModel } from './models/label.model';
 import { AuthRepository } from './repository/auth.repository';
 import { ChamaaiRepository } from './repository/chamaai.repository';
 import { ChatRepository } from './repository/chat.repository';
 import { ChatwootRepository } from './repository/chatwoot.repository';
 import { ContactRepository } from './repository/contact.repository';
+import { IntegrationRepository } from './repository/integration.repository';
+import { LabelRepository } from './repository/label.repository';
 import { MessageRepository } from './repository/message.repository';
 import { MessageUpRepository } from './repository/messageUp.repository';
 import { ProxyRepository } from './repository/proxy.repository';
@@ -52,6 +57,7 @@ import { AuthService } from './services/auth.service';
 import { CacheService } from './services/cache.service';
 import { ChamaaiService } from './services/chamaai.service';
 import { ChatwootService } from './services/chatwoot.service';
+import { IntegrationService } from './services/integration.service';
 import { WAMonitoringService } from './services/monitor.service';
 import { ProxyService } from './services/proxy.service';
 import { RabbitmqService } from './services/rabbitmq.service';
@@ -74,9 +80,11 @@ const proxyRepository = new ProxyRepository(ProxyModel, configService);
 const chamaaiRepository = new ChamaaiRepository(ChamaaiModel, configService);
 const rabbitmqRepository = new RabbitmqRepository(RabbitmqModel, configService);
 const sqsRepository = new SqsRepository(SqsModel, configService);
+const integrationRepository = new IntegrationRepository(IntegrationModel, configService);
 const chatwootRepository = new ChatwootRepository(ChatwootModel, configService);
 const settingsRepository = new SettingsRepository(SettingsModel, configService);
-const authRepository = new AuthRepository(AuthModel, configService);
+const authRepository = new AuthRepository(AuthModel, IntegrationModel, configService);
+const labelRepository = new LabelRepository(LabelModel, configService);
 
 export const repository = new RepositoryBroker(
   messageRepository,
@@ -92,7 +100,9 @@ export const repository = new RepositoryBroker(
   typebotRepository,
   proxyRepository,
   chamaaiRepository,
+  integrationRepository,
   authRepository,
+  labelRepository,
   configService,
   dbserver?.getClient(),
 );
@@ -111,7 +121,7 @@ export const typebotController = new TypebotController(typebotService);
 
 const webhookService = new WebhookService(waMonitor);
 
-export const webhookController = new WebhookController(webhookService);
+export const webhookController = new WebhookController(webhookService, waMonitor);
 
 const websocketService = new WebsocketService(waMonitor);
 
@@ -132,6 +142,8 @@ export const rabbitmqController = new RabbitmqController(rabbitmqService);
 const sqsService = new SqsService(waMonitor);
 
 export const sqsController = new SqsController(sqsService);
+
+const integrationService = new IntegrationService(waMonitor);
 
 const chatwootService = new ChatwootService(waMonitor, configService, repository, chatwootCache);
 
@@ -154,11 +166,13 @@ export const instanceController = new InstanceController(
   rabbitmqService,
   sqsService,
   typebotService,
+  integrationService,
   cache,
   chatwootCache,
 );
 export const sendMessageController = new SendMessageController(waMonitor);
 export const chatController = new ChatController(waMonitor);
 export const groupController = new GroupController(waMonitor);
+export const labelController = new LabelController(waMonitor);
 
 logger.info('Module - ON');